Create Clear-TempFolders.ps1
Typically, the path to the user's TEMP folder in the registry contains a relative path that refers to the USERPROFILE system variable.
When the registry value is read, the runtime automatically places the running process owner's profile path in the USERPROFILE variable.
Therefore, to get the correct path to the user's TEMP folder, the registry value referencing USERPROFILE must be corrected by replacing the process owner's profile path with the user's profile path.