Update Procedure Set Black Desktop Background.xml
The procedure (1) sets the desktop background for all users, and (2) sets the desktop background every time a user logs in.
To set the background (1), the AP runs script that loads users' registry hives, changes them, and triggers applying the settings made. To run the script at logon (2), the AP creates a link to the script in the Startup menu for all users.

However, there is a pitfall in (1). The AP runs the script under the System account, which is allowed to make registry changes for all users.
But the script triggers applying changes for the principal of the running process, which is the System. Thus, the script doesn't immediately affect the logged-in users. It needs the users to re-logon to see the changes.

To solve this, the AP creates scheduled tasks for each of the logged in users to run the script. That does the trick: the scheduled script runs under the account of logged in user and triggers changes for its principal in 30 seconds (The changed applied here are user-specific as the script has no access to other users' profiles).